--- .firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "tenant-file-fc6de"
-  }
-}
