--- functions/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "functions",
-  "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "build": "tsc",
-    "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "@google-cloud/storage": "^4.5.0",
-    "@types/node-fetch": "^2.5.5",
-    "ext-name": "^5.0.0",
-    "firebase": "^7.9.1",
-    "firebase-admin": "^8.6.0",
-    "firebase-functions": "^3.3.0",
-    "node-fetch": "^2.6.0",
-    "twilio": "^3.39.4"
-  },
-  "devDependencies": {
-    "tslint": "^5.12.0",
-    "typescript": "^3.2.2",
-    "firebase-functions-test": "^0.1.6"
-  },
-  "private": true
-}
