--- storage.rules
@@ -1,7 +1,0 @@
-service firebase.storage {
-  match /b/{bucket}/o {
-    match /{allPaths=**} {
-      allow read, write: if request.auth!=null;
-    }
-  }
-}
