+++ tenant-file/functions/package.json
@@ -14,8 +14,13 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "@google-cloud/storage": "^4.5.0",
+    "@types/node-fetch": "^2.5.5",
+    "ext-name": "^5.0.0",
+    "firebase": "^7.9.1",
     "firebase-admin": "^8.6.0",
     "firebase-functions": "^3.3.0",
+    "node-fetch": "^2.6.0",
     "twilio": "^3.39.4"
   },
   "devDependencies": {
