+++ tenant-file/functions/src/global.d.ts
